add steam native

+++ gaming.sh
@@ -13,38 +13,51 @@ flatpak install -y flathub org.freedesktop.Platform.VulkanLayer.MangoHud//23.08
 # Install Gamescope
 flatpak install -y flathub org.freedesktop.Platform.VulkanLayer.gamescope//23.08
 
+if [ ${STEAM_VERSION} = "native" ]; then
+    # Install Gamemode
+    pacman -S --noconfirm gamemode lib32-gamemode
+
+    # Install Gamescope
+    pacman -S --noconfirm gamescope
+
+    # Install MangoHud
+    pacman -S --noconfirm mangohud lib32-mangohud
+
+    # Configure MangoHud
+    mkdir -p /home/${NEW_USER}/.config/MangoHud
+    curl -sSL https://raw.githubusercontent.com/gjpin/arch-linux/main/configs/mangohud/MangoHud.conf -o /home/${NEW_USER}/.config/MangoHud/MangoHud.conf
+fi
+
 ################################################
-##### Steam (Flatpak)
+##### Steam
 ################################################
 
-# Install Steam
-flatpak install -y flathub com.valvesoftware.Steam
+if [ ${STEAM_VERSION} = "native" ]; then
+    # Install Steam
+    pacman -S --noconfirm steam
 
-# Create directory for Steam games
-mkdir -p /home/${NEW_USER}/Games/Steam
+    # Make sure correct driver is installed
+    if pacman -Qs lib32-amdvlk > /dev/null; then
+        pacman -Rs --noconfirm lib32-amdvlk
+        pacman -S --noconfirm lib32-vulkan-radeon
+    fi
+elif [ ${STEAM_VERSION} = "flatpak" ]; then
+    # Install Steam
+    flatpak install -y flathub com.valvesoftware.Steam
 
-# Import Flatpak overrides
-curl https://raw.githubusercontent.com/gjpin/arch-linux/main/configs/flatpak/com.valvesoftware.Steam -o /home/${NEW_USER}/.local/share/flatpak/overrides/com.valvesoftware.Steam
+    # Import Flatpak overrides
+    curl https://raw.githubusercontent.com/gjpin/arch-linux/main/configs/flatpak/com.valvesoftware.Steam -o /home/${NEW_USER}/.local/share/flatpak/overrides/com.valvesoftware.Steam
+
+    # Configure MangoHud for Steam
+    mkdir -p /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/config/MangoHud
+    curl -sSL https://raw.githubusercontent.com/gjpin/arch-linux/main/configs/mangohud/MangoHud.conf -o /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/config/MangoHud/MangoHud.conf
+fi
 
 # Steam controllers udev rules
 curl -sSL https://raw.githubusercontent.com/ValveSoftware/steam-devices/master/60-steam-input.rules -o /etc/udev/rules.d/60-steam-input.rules
 
-# Configure MangoHud for Steam
-mkdir -p /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/config/MangoHud
-tee /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/config/MangoHud/MangoHud.conf << EOF
-legacy_layout=0
-horizontal
-gpu_stats
-cpu_stats
-ram
-fps
-frametime=0
-hud_no_margin
-table_columns=14
-frame_timing=1
-engine_version
-vulkan_driver
-EOF
+# Create directory for Steam games
+mkdir -p /home/${NEW_USER}/Games/Steam
 
 ################################################
 ##### Heroic Games Launcher (Flatpak)
@@ -64,20 +77,7 @@ curl https://raw.githubusercontent.com/gjpin/arch-linux/main/configs/flatpak/com
 
 # Configure MangoHud for Heroic
 mkdir -p /home/${NEW_USER}/.var/app/com.heroicgameslauncher.hgl/config/MangoHud
-tee /home/${NEW_USER}/.var/app/com.heroicgameslauncher.hgl/config/MangoHud/MangoHud.conf << EOF
-legacy_layout=0
-horizontal
-gpu_stats
-cpu_stats
-ram
-fps
-frametime=0
-hud_no_margin
-table_columns=14
-frame_timing=1
-engine_version
-vulkan_driver
-EOF
+curl -sSL https://raw.githubusercontent.com/gjpin/arch-linux/main/configs/mangohud/MangoHud.conf -o /home/${NEW_USER}/.var/app/com.heroicgameslauncher.hgl/config/MangoHud/MangoHud.conf
 
 ################################################
 ##### Sunshine (native)
@@ -110,13 +110,13 @@ fi
 setcap cap_sys_admin+p $(readlink -f /usr/bin/sunshine)
 
 # Allow Sunshine in firewall
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="48010" protocol="tcp" accept log prefix="Sunshine - RTSP TCP"'
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="48010" protocol="udp" accept log prefix="Sunshine - RTSP UDP"'
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="47998" protocol="udp" accept log prefix="Sunshine - Video"'
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="48000" protocol="udp" accept log prefix="Sunshine - Audio"'
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="48010" protocol="tcp" accept log prefix="Sunshine - RTSP TCP"'
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="48010" protocol="udp" accept log prefix="Sunshine - RTSP UDP"'
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="47998" protocol="udp" accept log prefix="Sunshine - Video"'
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="48000" protocol="udp" accept log prefix="Sunshine - Audio"'
 
 ################################################
-##### ALVR (flatpak)
+##### ALVR
 ################################################
 
 # References:
@@ -124,44 +124,34 @@ firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port p
 # https://github.com/alvr-org/ALVR/wiki/Installation-guide#portable-targz
 # https://github.com/alvr-org/ALVR/tree/master/alvr/xtask/firewall
 # https://github.com/alvr-org/ALVR/wiki/Flatpak
+# https://github.com/alvr-org/ALVR/wiki/Installation-guide
 
-# Download ALVR flatpak
-curl https://github.com/alvr-org/ALVR/releases/latest/download/com.valvesoftware.Steam.Utility.alvr.flatpak -L -O
+if [ ${STEAM_VERSION} = "native" ]; then
+    # Install ALVR
+    sudo -u ${NEW_USER} paru -S --noconfirm alvr-bin
+elif [ ${STEAM_VERSION} = "flatpak" ]; then
+    # Download ALVR flatpak
+    curl https://github.com/alvr-org/ALVR/releases/latest/download/com.valvesoftware.Steam.Utility.alvr.flatpak -L -O
 
-# Install ALVR
-flatpak install -y --bundle com.valvesoftware.Steam.Utility.alvr.flatpak
+    # Install ALVR
+    flatpak install -y --bundle com.valvesoftware.Steam.Utility.alvr.flatpak
 
-# Remove ALVR flatpak
-rm -f com.valvesoftware.Steam.Utility.alvr.flatpak
+    # Remove ALVR flatpak
+    rm -f com.valvesoftware.Steam.Utility.alvr.flatpak
 
-# Automatic Audio & Microphone setup
-curl https://raw.githubusercontent.com/alvr-org/ALVR/master/alvr/xtask/flatpak/audio-flatpak-setup.sh -o /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/audio-flatpak-setup.sh
-chmod +x /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/audio-flatpak-setup.sh
+    # Automatic Audio & Microphone setup
+    curl https://raw.githubusercontent.com/alvr-org/ALVR/master/alvr/xtask/flatpak/audio-flatpak-setup.sh -o /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/audio-flatpak-setup.sh
+    chmod +x /home/${NEW_USER}/.var/app/com.valvesoftware.Steam/audio-flatpak-setup.sh
 
-# Create ALVR shortcut
-tee /home/${NEW_USER}/.local/share/applications/alvr.desktop << EOF
-[Desktop Entry]
-Version=1.0
-Type=Application
-Name=ALVR
-GenericName=Game
-Comment=ALVR is an open source remote VR display which allows playing SteamVR games on a standalone headset such as Gear VR or Oculus Go/Quest.
-Exec=/usr/bin/flatpak run --command=alvr_dashboard com.valvesoftware.Steam
-Icon=alvr
-Categories=Game;
-StartupNotify=true
-PrefersNonDefaultGPU=true
-X-KDE-RunOnDiscreteGpu=true
-StartupWMClass=ALVR
-EOF
+    # Create ALVR shortcut
+    curl https://raw.githubusercontent.com/gjpin/arch-linux/main/configs/alvr/alvr-flatpak.desktop -o /home/${NEW_USER}/.local/share/applications/alvr.desktop
+
+    # Create ALVR dashboard alias
+    echo 'alias alvr="flatpak run --command=alvr_dashboard com.valvesoftware.Steam"' | tee /home/${NEW_USER}/.zshrc.d/alvr
+fi
 
 # Allow ALVR in firewall
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="9943" protocol="udp" accept log prefix="ALVR - discovery UDP"'
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="9944" protocol="udp" accept log prefix="ALVR - SteamVR UDP"'
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="9943" protocol="tcp" accept log prefix="ALVR - discovery TCP"'
-firewall-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" port port="9944" protocol="tcp" accept log prefix="ALVR - SteamVR TCP"'
-
-# Create ALVR dashboard alias
-tee /home/${NEW_USER}/.zshrc.d/alvr << 'EOF'
-alias alvr="flatpak run --command=alvr_dashboard com.valvesoftware.Steam"
-EOF
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="9943" protocol="udp" accept log prefix="ALVR - discovery UDP"'
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="9944" protocol="udp" accept log prefix="ALVR - SteamVR UDP"'
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="9943" protocol="tcp" accept log prefix="ALVR - discovery TCP"'
+firewall-offline-cmd --permanent --zone=block --add-rich-rule='rule family="ipv4" source address="10.100.100.0/24" port port="9944" protocol="tcp" accept log prefix="ALVR - SteamVR TCP"'
